# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PLRD Forum API",
    "summary": "Read and write endpoints for PLRD Forum, a LessWrong-style ATProto forum.",
    "description": "PLRD Forum stores all content as records in users' ATProto data repositories (site.standard.* lexicons). Read endpoints are public; writes require an ATProto OAuth session and create records in the signed-in user's own PDS. Pages also support markdown content negotiation via Accept: text/markdown, and an MCP server is published at /.well-known/mcp with tools get_frontpage, get_post, and list_concepts.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://plrd-forum.vercel.app" }],
  "paths": {
...
    "/api/posts": {
      "get": {
        "operationId": "listPosts",
        "summary": "List indexed posts (public JSON)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "number", "default": 25, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts ranked by recency with karma and comment counts",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      },
      "post": {
        "operationId": "createPost",
        "summary": "Create a post (authenticated)",
        "description": "Writes a site.standard.document record to the signed-in user's PDS; auto-creates their publication on first post.",
        "security": [{ "atprotoOAuth": [] }],
        "responses": { "200": { "description": "Post created" }, "401": { "description": "Not signed in" } }
      }
    },
...
  },
...
}